+++ tests/unit/virtual-auto-combo.test.ts
@@ -91,6 +91,61 @@ test("createVirtualAutoCombo includes configured web-session providers without a
   assert.ok(combo.autoConfig.candidatePool.includes("qwen-web"));
 });
 
+test("createVirtualAutoCombo excludes web-session providers with empty required token data", async () => {
+  await providersDb.createProviderConnection({
+    provider: "qwen-web",
+    authType: "apikey",
+    name: "Qwen Web Empty Session",
+    providerSpecificData: { token: "   " },
+    defaultModel: "qwen3-coder-plus",
+  });
+
+  const combo: VirtualComboResult = await virtualFactory.createVirtualAutoCombo("coding");
+
+  assert.equal(
+    combo.models.some((model) => model.providerId === "qwen-web"),
+    false,
+    "web-session providers with empty required token data must not be auto-combo candidates"
+  );
+  assert.equal(combo.autoConfig.candidatePool.includes("qwen-web"), false);
+});
+
+test("createVirtualAutoCombo excludes web-session providers with irrelevant providerSpecificData", async () => {
+  await providersDb.createProviderConnection({
+    provider: "chatgpt-web",
+    authType: "apikey",
+    name: "ChatGPT Web Invalid Session",
+    providerSpecificData: { unrelated: "value" },
+    defaultModel: "gpt-4o",
+  });
+
+  const combo: VirtualComboResult = await virtualFactory.createVirtualAutoCombo("coding");
+
+  assert.equal(
+    combo.models.some((model) => model.providerId === "chatgpt-web"),
+    false,
+    "web-session providers with irrelevant providerSpecificData must not be auto-combo candidates"
+  );
+  assert.equal(combo.autoConfig.candidatePool.includes("chatgpt-web"), false);
+});
+
+test("createVirtualAutoCombo includes cookie web-session providers with required cookie data", async () => {
+  await providersDb.createProviderConnection({
+    provider: "chatgpt-web",
+    authType: "apikey",
+    name: "ChatGPT Web Session",
+    providerSpecificData: { cookie: "__Secure-next-auth.session-token=chatgpt-session" },
+    defaultModel: "gpt-4o",
+  });
+
+  const combo: VirtualComboResult = await virtualFactory.createVirtualAutoCombo("coding");
+
+  const chatgptWeb = combo.models.find((model) => model.providerId === "chatgpt-web");
+  assert.ok(chatgptWeb, "cookie web-session providers with required cookie data should be candidates");
+  assert.equal(chatgptWeb.model, "chatgpt-web/gpt-4o");
+  assert.ok(combo.autoConfig.candidatePool.includes("chatgpt-web"));
+});
+
 test("createVirtualAutoCombo includes no-auth OpenCode Free without provider_connections rows", async () => {
   const combo: VirtualComboResult = await virtualFactory.createVirtualAutoCombo("fast");
